+++ example/lib/pages/path_finding/human_path_finding.dart
@@ -3,9 +3,7 @@ import 'package:example/pages/player/simple/human.dart';
 
 class HumanPathFinding extends HumanPlayer with PathFinding, TapGesture {
   HumanPathFinding({required Vector2 position}) : super(position: position) {
-    setupPathFinding(
-      pathLineStrokeWidth: 2,
-    );
+    setupPathFinding(pathLineStrokeWidth: 2);
   }
 
   @override