+++ lib/tiled/reader/tiled_asset_reader.dart
@@ -5,7 +5,6 @@ import 'package:tiledjsonreader/tiledjsonreader.dart';
 class TiledAssetReader extends TiledReader {
 // ignore: constant_identifier_names
   static const _ASSET_KEY_PREFIX = 'assets/images/';
-  static const ORIENTATION_SUPPORTED = 'orthogonal';
   final String asset;
   late TiledJsonReader _reader;
   @override